# src/app/lib/API/swagger.yaml
openapi: 3.0.3
info:
  title: LEGO Inventory API
  description: API
  version: 1.0.0
servers:
  - url: http://localhost:3000/api/v1
    description: API v1 Dev
components:
  schemas:
...
    PartSearchResult:
      type: object
      properties:
        id:
          type: string
          description: Part ID
        name:
          type: string
          description: Part name
      required:
        - id
        - name
    SetSearchResult:
      type: object
      properties:
        id:
          type: string
          description: Set ID
        name:
          type: string
          description: Set name
      required:
        - id
        - name
    MinifigSearchResult:
      type: object
      properties:
        id:
          type: string
          description: Minifig ID
        name:
          type: string
          description: Minifig name
      required:
        - id
        - name
...
    Error:
      type: object
      properties:
        code:
          type: integer
          description: Error code
        message:
          type: string
          description: Error message
      required:
        - code
        - message
paths:
...
  /search/part/{searchTerm}:
    get:
      tags:
        - Search
      summary: Search for LEGO parts
      description: Searches for LEGO parts by name or ID
      operationId: searchParts
      parameters:
        - name: searchTerm
          in: path
          description: Search term
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      $ref: "#/components/schemas/PartSearchResult"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /search/set/{searchTerm}:
    get:
      tags:
        - Search
      summary: Search for LEGO sets
      description: Searches for LEGO sets by name or set number
      operationId: searchSets
      parameters:
        - name: searchTerm
          in: path
          description: Search term
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      $ref: "#/components/schemas/SetSearchResult"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /search/minifig/{searchTerm}:
    get:
      tags:
        - Search
      summary: Search for LEGO minifigs
      description: Searches for LEGO minifigs by name or ID
      operationId: searchMinifigs
      parameters:
        - name: searchTerm
          in: path
          description: Search term
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      $ref: "#/components/schemas/MinifigSearchResult"
                  note:
                    type: string
                  aborted:
                    type: boolean
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...